\name{pcfinhom-package}
\alias{pcfinhom-package}
\alias{pcfinhom}
\docType{package}
\title{
\packageTitle{pcfinhom}
}
\description{
\packageDescription{pcfinhom}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pcfinhom}
\packageIndices{pcfinhom}

This package accompanies the paper: .
It provides estimators for the non-parametric metrics of
second order interactions in point processes, when the intensity function 
is not uniform. It makes heavy use of the spatstat package for point
pattern analysis.
}
\author{
\packageAuthor{pcfinhom}

Maintainer: \packageMaintainer{pcfinhom}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{
}
